backend/requirements.txt
# API Backend
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-multipart>=0.0.9

# Data Processing & ORM
sqlalchemy>=2.0.31
pandas>=2.2.2
openpyxl>=3.1.5

# Utilities
python-dotenv>=1.0.1
